"First loggin form local database"

+++ app/src/main/java/com/example/ecocyam/localDatabase/DatabaseHelper.java
@@ -2,6 +2,7 @@ package com.example.ecocyam.localDatabase;
 
 import android.content.ContentValues;
 import android.content.Context;
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
@@ -10,6 +11,7 @@ import androidx.annotation.Nullable;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     public static final String DATABASE_NAME = "ecocyam.db";
+    public static final int VERSION = 2;
 
     public static final String TABLE_NAME = "user_table";
     public static final String COL_ID = "ID";
@@ -17,7 +19,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     public static final String COL_PASSWORD = "PASSWORD";
 
     public DatabaseHelper(@Nullable Context context) {
-        super(context, DATABASE_NAME, null, 1);
+        super(context, DATABASE_NAME, null, VERSION);
 
     }
 
@@ -30,6 +32,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        if (newVersion>oldVersion)
         db.execSQL("DROP TABLE IF EXISTS "+TABLE_NAME);
         onCreate(db);
     }
@@ -48,5 +51,51 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         else
             return true;
 
+    }
+
+    public Cursor get_data(){
+        SQLiteDatabase db=this.getWritableDatabase();
+        Cursor cursor = db.rawQuery("SELECT * FROM "+TABLE_NAME,null);
+        return cursor;
+    }
+
+    public boolean updateUser(String id,String user, String password){
+        SQLiteDatabase db=this.getWritableDatabase();
+        ContentValues contentValues = new ContentValues();
+
+        contentValues.put(COL_ID,id);
+        contentValues.put(COL_NAME,user);
+        contentValues.put(COL_PASSWORD,password);
+
+        db.update(TABLE_NAME,contentValues,"ID = ?",new String[] {id});
+        return true;
+    }
+
+    public int deleteUser(String id){
+        SQLiteDatabase db=this.getWritableDatabase();
+        return db.delete(TABLE_NAME,"ID = ?", new String[]{id});
+
+    }
+
+    public boolean isUserExist(String userName) {
+        SQLiteDatabase db=this.getWritableDatabase();
+
+        Cursor cursor = db.rawQuery("SELECT * FROM "+TABLE_NAME + " WHERE NAME=?",new String[] {userName});
+        if (cursor.getCount()==0){
+            return true;
+        }
+        //user already exists
+        return false;
+    }
+
+    public boolean testConnectionInfo(String userName, String password) {
+        SQLiteDatabase db=this.getWritableDatabase();
+
+        Cursor cursor = db.rawQuery("SELECT * FROM "+TABLE_NAME + " WHERE NAME=? AND PASSWORD=?",new String[] {userName,password});
+        if (cursor.getCount()==0){
+            return false;
+        }
+        //user is existed
+        return true;
     }
 }
